requirements.txt
pymongo==3.10.1
pytest==6.1.1
pyyaml==5.3
requests==2.22.0
dnspython==1.16.0
numpy==1.18.1
termcolor==1.1.0
black==20.8b1
